--- src/components/Table/components/index.ts
@@ -1,7 +1,0 @@
-export { defaultRenderFn, headerRenderFn, rowClickRenderFn, rowLinkRenderFn } from "./cell";
-export type { GridCellContent, RenderCellFn } from "./cell";
-export * from "./CollapseToggle";
-export { EditColumnsButton } from "./EditColumnsButton";
-export { Row } from "./Row";
-export * from "./SelectToggle";
-export * from "./SortHeader";
