--- src/components/Table/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from "./useColumns";
-export * from "./useSetupColumnSizes";
-export * from "./useSortState";
