+++ src/components/Table/index.ts
@@ -1,17 +1,37 @@
+export {
+  defaultRenderFn,
+  headerRenderFn,
+  rowClickRenderFn,
+  rowLinkRenderFn,
+} from "src/components/Table/components/cell";
+export type { GridCellContent, RenderCellFn } from "src/components/Table/components/cell";
+export * from "src/components/Table/components/CollapseToggle";
+export { EditColumnsButton } from "src/components/Table/components/EditColumnsButton";
+export { Row } from "src/components/Table/components/Row";
+export type { GridDataRow, GridRowKind } from "src/components/Table/components/Row";
 export * from "src/components/Table/components/SelectToggle";
+export * from "src/components/Table/components/SortHeader";
 export { SortHeader } from "src/components/Table/components/SortHeader";
-export { useColumns } from "src/components/Table/hooks/useColumns";
+export { useGridTableApi } from "src/components/Table/GridTableApi";
+export type { GridTableApi } from "src/components/Table/GridTableApi";
+export * from "src/components/Table/hooks/useColumns";
+export * from "src/components/Table/hooks/useSetupColumnSizes";
+export * from "src/components/Table/hooks/useSortState";
+export { cardStyle, condensedStyle, defaultStyle, getTableStyles } from "src/components/Table/TableStyles";
+export type { GridStyle, RowStyle, RowStyles } from "src/components/Table/TableStyles";
+export * from "src/components/Table/types";
 export * from "src/components/Table/utils/columns";
+export { createRowLookup } from "src/components/Table/utils/GridRowLookup";
 export type { GridRowLookup } from "src/components/Table/utils/GridRowLookup";
+export * from "src/components/Table/utils/GridSortContext";
 export { GridSortContext } from "src/components/Table/utils/GridSortContext";
+export type { GridSortContextProps } from "src/components/Table/utils/GridSortContext";
 export { RowState, RowStateContext } from "src/components/Table/utils/RowState";
+export type { SelectedState } from "src/components/Table/utils/RowState";
 export { simpleDataRows, simpleHeader } from "src/components/Table/utils/simpleHelpers";
 export type { SimpleHeaderAndData } from "src/components/Table/utils/simpleHelpers";
-export * from "./components";
+export * from "src/components/Table/utils/sortRows";
+export * from "src/components/Table/utils/utils";
+export * from "src/components/Table/utils/visitor";
 export { GridTable, setDefaultStyle, setGridTableDefaults } from "./GridTable";
 export type { GridSortConfig, GridTableDefaults, GridTableProps, setRunningInJest } from "./GridTable";
-export { useGridTableApi } from "./GridTableApi";
-export type { GridTableApi } from "./GridTableApi";
-export * from "./hooks";
-export * from "./types";
-export * from "./utils";
