--- src/components/Table/utils/index.ts
@@ -1,14 +1,0 @@
-export * from "./columns";
-export { createRowLookup } from "./GridRowLookup";
-export type { GridRowLookup } from "./GridRowLookup";
-export * from "./GridSortContext";
-export type { GridSortContextProps } from "./GridSortContext";
-export { RowState, RowStateContext } from "./RowState";
-export type { SelectedState } from "./RowState";
-export { simpleDataRows } from "./simpleHelpers";
-export type { SimpleHeaderAndData } from "./simpleHelpers";
-export * from "./sortRows";
-export { cardStyle, condensedStyle, defaultStyle, getTableStyles } from "./styles";
-export type { GridStyle, RowStyle, RowStyles } from "./styles";
-export * from "./utils";
-export * from "./visitor";
