+++ src/components/Table/utils/sortRows.ts
@@ -1,7 +1,8 @@
 import { ReactNode } from "react";
-import { GridCellContent } from "src/components/Table/components";
+import { GridCellContent } from "src/components/Table/components/cell";
+import { GridDataRow } from "src/components/Table/components/Row";
 import { SortOn, SortState } from "src/components/Table/hooks/useSortState";
-import { GridColumn, GridDataRow, Kinded, Pin } from "src/components/Table/types";
+import { GridColumn, Kinded, Pin } from "src/components/Table/types";
 import { applyRowFn } from "src/components/Table/utils/utils";
 
 // Returns a shallow copy of the `rows` parameter sorted based on `sortState`
