+++ src/components/Table/utils/utils.tsx
@@ -1,16 +1,10 @@
 import React, { ReactNode } from "react";
-import { GridCellContent } from "src/components/Table/components";
+import { GridCellContent } from "src/components/Table/components/cell";
+import { GridDataRow } from "src/components/Table/components/Row";
 import { SortHeader } from "src/components/Table/components/SortHeader";
 import { GridTableApi } from "src/components/Table/GridTableApi";
-import {
-  GridCellAlignment,
-  GridColumn,
-  GridDataRow,
-  Kinded,
-  ParentChildrenTuple,
-  RenderAs,
-} from "src/components/Table/types";
-import { GridStyle, RowStyle } from "src/components/Table/utils/styles";
+import { GridStyle, RowStyle } from "src/components/Table/TableStyles";
+import { GridCellAlignment, GridColumn, Kinded, RenderAs } from "src/components/Table/types";
 import { Css, Properties } from "src/Css";
 import { getButtonOrLink } from "src/utils/getInteractiveElement";
 
@@ -91,52 +85,6 @@ export function applyRowFn<R extends Kinded>(
   } else {
     return maybeContent;
   }
-}
-
-/**
- * Filters rows given a client-side text `filter.
- *
- * Ensures parent rows remain in the list if any children match the filter.
- *
- * We return a copy of `[Parent, [Child]]` tuples so that we don't modify the `GridDataRow.children`.
- */
-export function filterRows<R extends Kinded>(
-  api: GridTableApi<R>,
-  columns: GridColumn<R>[],
-  rows: GridDataRow<R>[],
-  filter: string | undefined,
-): ParentChildrenTuple<R>[] {
-  // Make a functions to do recursion
-  function acceptAll(acc: ParentChildrenTuple<R>[], row: GridDataRow<R>): ParentChildrenTuple<R>[] {
-    return acc.concat([[row, row.children?.reduce(acceptAll, []) ?? []]]);
-  }
-
-  function filterFn(acc: ParentChildrenTuple<R>[], row: GridDataRow<R>): ParentChildrenTuple<R>[] {
-    // Break up "foo bar" into `[foo, bar]` and a row must match both `foo` and `bar`
-    const filters = (filter && filter.split(/ +/)) || [];
-    const matches =
-      row.kind === "header" ||
-      row.kind === "totals" ||
-      filters.length === 0 ||
-      filters.every((f) =>
-        columns.map((c) => applyRowFn(c, row, api, 0)).some((maybeContent) => matchesFilter(maybeContent, f)),
-      );
-    if (matches) {
-      return acc.concat([[row, row.children?.reduce(acceptAll, []) ?? []]]);
-    } else {
-      const matchedChildren = row.children?.reduce(filterFn, []) ?? [];
-      if (
-        matchedChildren.length > 0 ||
-        typeof row.pin === "string" ||
-        (row.pin !== undefined && row.pin.filter !== true)
-      ) {
-        return acc.concat([[row, matchedChildren]]);
-      } else {
-        return acc;
-      }
-    }
-  }
-  return rows.reduce(filterFn, []);
 }
 
 export const ASC = "ASC" as const;
