+++ src/components/Table/utils/visitor.ts
@@ -1,4 +1,4 @@
-import { GridDataRow } from "src/components/Table/types";
+import { GridDataRow } from "src/components/Table/components/Row";
 
 export function visit(rows: GridDataRow<any>[], fn: (row: GridDataRow<any>) => void): void {
   const todo = [...rows];
